Create convDFT.cu
This is the final implementation with streams to calculate the convolution of two input signals. It takes their DFT, takes the dot product of the resultant DFTs of the input signal and then takes the IDFT of the resultant product. The time domain resultant signal thus obtained is the convolution of the two input signals. The results have been verified using MATLAB.